pyproject.toml
[tool.poetry]
name = "bigquery-subscription"
version = "0.1.0"
description = ""
authors = ["MtkN1 <51289448+MtkN1@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
websockets = "^12.0"
google-cloud-pubsub = "^2.19.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
